refactor(repo-lint): split control-flow nesting scan from CLI checker
Extract visitors and collectControlFlowNestingDepthViolationsFromUnit into
tool/control_flow_nesting_depth_scan.dart; keep check_control_flow_nesting_depth
as thin orchestration + export (mirrors disallowed-AST rules/visitor split).

Tests import the scan module directly. Update SPEC/program/control-flow-nesting-depth.md
and repo-lint artifact table. Refs #1912.
